Adds pagination_list_item_attrs to right ellipsis list item.
This was missing from the #337 PR.

This commit adds the `pagination_list_item_attrs` option to the right
ellipsis list item.
The test to cover this only verified that the first list child had
the assigned options.
I've changed it to verify all list items, and added opts to trigger
ellipsis rendering.